executor-py/pyproject.toml
[tool.poetry]
name = "indexify-executor"
version = "0.0.1"
description = "Indexify Executor CLI"
authors = ["Diptanu Gon Choudhury <diptanu@tensorlake.ai>"]
readme = "README.md"

[tool.poetry.scripts]
indexify-executor = "indexify_executor.main:typer_app"

[tool.poetry.dependencies]
python = "^3.9"
typer = { extras = ["all"], version = "^0.9.0" }
pydantic = "^2.6.1"
nanoid = "^2.0.0"
rich = "^13.7.1"
pyyaml = "^6.0.1"
httpx = "^0"
indexify = "0.0.44"

[tool.poetry.dev-dependencies]
syrupy = "^4.0.0"
pytest = ">7.0.0"

[tool.poetry.group.dev.dependencies]
asynctest = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
